+++ src/tokenizers/words/index.js
@@ -1,0 +1,2 @@
+import naive from './naive';
+export default naive;